Lab/requirements.txt
# requirements.txt

python-telegram-bot==20.3
httpx==0.24.1
Pillow
pytesseract
gTTS
markovify
deep-translator
python-dotenv
SpeechRecognition
imageio-ffmpeg
easyocr
